requirements.txt
pyzmq==19.0.2
numpy==1.18.5
Pillow==8.0.1
opencv-python==4.4.0.46
flask
